requirements.txt
absl-py==2.4.0
aiofiles==25.1.0
array_record==0.8.3
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
chex==0.1.91
cloudpickle==3.1.2
dataclasses-json==0.6.7
etils==1.13.0
flatbuffers==25.12.19
flax==0.12.3
fsspec==2026.1.0
gast==0.7.0
google-pasta==0.2.0
grain==0.2.15
grpcio==1.76.0
h5py==3.15.1
humanize==4.15.0
idna==3.11
importlib_resources==6.5.2
jax==0.9.0
jax-cuda13-pjrt==0.9.0
jax-cuda13-plugin==0.9.0
jaxlib==0.9.0
jaxtyping==0.3.6
joblib==1.5.3
keras==3.13.1
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
ml_dtypes==0.5.4
msgpack==1.1.2
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
numpy==2.3.5
nvidia-cublas==13.2.1.1
nvidia-cuda-cccl==13.1.115
nvidia-cuda-crt==13.1.115
nvidia-cuda-cupti==13.1.115
nvidia-cuda-nvcc==13.1.115
nvidia-cuda-nvrtc==13.1.115
nvidia-cuda-runtime==13.1.80
nvidia-cudnn-cu13==9.18.1.3
nvidia-cufft==12.1.0.78
nvidia-cusolver==12.0.9.81
nvidia-cusparse==12.7.3.1
nvidia-nccl-cu13==2.29.2
nvidia-nvjitlink==13.1.115
nvidia-nvshmem-cu13==3.5.19
nvidia-nvvm==13.1.115
opt_einsum==3.4.0
optax==0.2.6
optree==0.18.0
orbax-checkpoint==0.11.32
orbax-export==0.0.8
packaging==26.0
pandas==3.0.0
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
protobuf==6.33.5
psutil==7.2.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
simplejson==3.20.2
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.20.0
tensorstore==0.1.80
termcolor==3.3.0
threadpoolctl==3.6.0
toolz==1.1.0
tqdm==4.67.1
treescope==0.1.10
typing-inspect==0.9.0
typing_extensions==4.15.0
urllib3==2.6.3
wadler_lindig==0.1.7
Werkzeug==3.1.5
wheel==0.46.3
wrapt==2.0.1
zipp==3.23.0
